keep all psam rows
Plink assumes that .psam, .pvar and .pgen files will all have the
same prefix, so it's better to keep them as a set even if the psam
file is identical for each chromosome. Changing the source of the
primary key to file_path allows us to have all files in the table.